requirements.txt
PySide6
requests
keyboard
Pillow
docx
openpyxl
pandas
pymupdf
pywin32
markdown
